+++ ryzom/client/src/interface_v3/interface_manager.cpp
@@ -2208,7 +2208,7 @@ void CInterfaceManager::drawViews(NL3D::UCamera camera)
 	NLGUI::CDBManager::getInstance()->flushObserverCalls();
 
 	// Update Player characteristics (for Item carac requirement Redifying)
-	nlctassert(CHARACTERISTICS::NUM_CHARACTERISTICS==8);
+	nlctassert((int)CHARACTERISTICS::NUM_CHARACTERISTICS==8);
 	for (uint i=0; i< (int)CHARACTERISTICS::NUM_CHARACTERISTICS; ++i)
 	{
 		if (!_CurrentPlayerCharacLeaf[i])
